backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-core
chromadb
pypdf
requests
python-dotenv
pydantic

